Converted priority from long to int to save a little memory per actively downloading torrent
git-svn-id: http://svn.vuze.com/public/client@8432 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core3/peer/PEPiece.java
@@ -101,13 +101,13 @@ public interface PEPiece
 	public void			setReservedBy(String peer);
 
 	/**
-	 * @return long ResumePriority (startPriority + resuming adjustments)
+	 * @return int ResumePriority (startPriority + resuming adjustments)
 	 */
-	public long			getResumePriority();
+	public int			getResumePriority();
 	/**
 	 * @param p the Resume Priority to set, for display purposes
 	 */
-	public void			setResumePriority(long p);
+	public void			setResumePriority(int p);
 
 	public String[] 	getWriters();
 	public void			setWritten(PEPeer peer, int blockNumber);
